Fixes #540.
The patch was not complete, it did not address chanid caching,
so it wouldn't work as is. But fixing this wasn't difficult..



git-svn-id: http://svn.mythtv.org/svn/trunk@7720 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/eithelper.h
@@ -35,7 +35,7 @@ class EITHelper : public QObject
 
     QListList_Events  eitList;      ///< Event Information Tables List
     mutable QMutex    eitList_lock; ///< EIT List lock
-    mutable QMap<uint,uint> srv_to_chanid;
+    mutable QMap<unsigned long long, uint> srv_to_chanid;
 
     /// Maximum number of DB inserts per ProcessEvents call.
     static const uint kChunkSize;
